fix windows compiling

+++ trunk/desmume/src/windows/oamView.cpp
@@ -224,7 +224,7 @@ LRESULT OamView_OnPaint(HWND hwnd, oamview_struct *win, WPARAM wParam, LPARAM lP
         for(i = 0; i < 192; ++i)
         {
 			copy.currLine = i;
-             copy.spriteRender((u8*)(bitmap + i*256), bitmap_alpha + i*256, type + i*256, prio + i*256);
+             copy.spriteRender((u16*)(bitmap + i*256), bitmap_alpha + i*256, type + i*256, prio + i*256);
         }
 
 		u32 width = dimm_int[(oam->attr1>>14)][(oam->attr0>>14)][0];
